text_describer/requirements.txt
google-generativeai
Pillow
# Requirements untuk Text Describer Flet Version
# Instalasi: pip install -r requirements.txt

# Framework UI Modern
flet>=0.21.0

# AI dan Machine Learning
google-generativeai>=0.3.0

# Utilities
Pillow>=10.0.0
requests>=2.31.0